requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
numpy==1.18.1
oauthlib==3.1.0
paho-mqtt==1.5.0
pandas==1.0.0
pkg-resources==0.0.0
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
query-string==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.7
zope.interface==5.1.0
